Issue 409: Implement message when there is no information.
When the tab section array list is empty, no content is displayed.
The component html template code is calling a function to get the section array list.

I want to avoid calling that function multiple times.
Get the section array list as a `sections` variable via an `ngIf`.
The `ngFor` can then loop over the `sections` variable.

The `sections?.length` is now being tested in the calling function so that I can return either a full object on nothing on an empty array.
This helps avoid the need for performing a `?.length` inside the html template code.
The `ngIf` does not handle assigning a value and operating on a member of that property very well.

That is to say this is not possible while passing sections as an array as far as I can tell:
```html
<div *ngIf="getSectionsToShow(tab, individual)?.length as sections ; else no_content">
```
The `?.length` would end up being added as sections.
Adding parenthesis around this also does not work:
```html
<div *ngIf="(getSectionsToShow(tab, individual) as sections)?.length ; else no_content">
```