410: Manually override pagination colors to enforce color contrast ratio.
This handles additional areas missed by this commit 59c5735d447652124525ae8129fc603a48688a54.

The `::deep` is necessary for the dialog because the ngb-pagination is in a different scope.

The global link colors use too light of a blue according to the scanners.
Change the color to a darker color.
The hover link is made darker to preserve the behavior of being darker on hover.